refactor: Move dojo nodes to neurochain
