+++ nodes/categories.py
@@ -30,6 +30,8 @@ AUDIO_CAT = f"{NEUROCHAIN_CAT}/🔊 Audio"
 
 CLEARML_CAT = f"{NEUROCHAIN_CAT}/ClearML"
 
+DOJO_CAT = f"{NEUROCHAIN_CAT}/⛩️ Dojo"
+
 EVALUATION_CAT = f"{NEUROCHAIN_CAT}/🔎 Evaluation"
 
 EMBEDDINGS_CAT = f"{NEUROCHAIN_CAT}/🔗 Embeddings"
@@ -58,5 +60,3 @@ NEUROCHAIN_UTILS_CAT = f"{NEUROCHAIN_CAT}/🛠️ Utils"
 ANY_CAT = f"{NEUROCHAIN_UTILS_CAT}/Any"
 DICTS_CAT = f"{NEUROCHAIN_UTILS_CAT}/Dicts"
 LISTS_CAT = f"{NEUROCHAIN_UTILS_CAT}/Lists"
-
-DOJO_CAT = f"{NAME}/⛩️ Dojo"
